backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "tuome",email = "tonni@tuta.io"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask (>=3.1.0,<4.0.0)",
    "google-generativeai (>=0.8.4,<0.9.0)",
    "python-dotenv (>=1.0.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"

